drm/imagination: Split up render job submission
Render jobs represent two seperate GPU jobs, geometry and fragment. Previously
the submission code would attempt to process and submit them in parallel as
much as possible. This lead to a very complicated processing function with
complex error handling, which was in practice broken and would cause
numerous use-after-free issues if the geometry job submission succeeded but
the fragment job submission failed.

This commit breaks up the render job submission so that geometry and fragment
jobs are processed and submitted seperately. This greatly simplifies the error
handling and should eliminate the use-after-free issues.

Signed-off-by: Sarah Walker <sarah.walker@imgtec.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>