drm/imagination: Job submission uAPI refactor
The job submit uAPI is a bit cumbersome, and is likely to become even
harder to apprehend once we throw advanced concepts like partial render
to the mix.

This is an attempt at simplifying things a bit. Now jobs are submitted
independently to the geometry and fragment queues instead of being
submitted as a mega job taking geometry and fragment commands. In order
to limit the number of userspace <-> kernelspace round trips, we turn the
the SUBMIT_JOB ioctl into a SUBMIT_JOBS one taking an array of job. This
way we can pass both the geometry and fragment commands in a single step,
but we also prepare for furture potential optimizations by allowing
compute, transfer and render jobs to be submitted in a single call.

Things are still submitted in-order, allowing one to wait for fences
returned by jobs placed earlier in the job array. With the consolidate
flag, we can also accumulate fences of several jobs in a single
syncobj point, making the NULL job concept redundant.

The implicit dependency between the geometry and fragment jobs now has
to be explicitly defined (easily done with an internal syncobj signaled
by the geometry job and waited-on by the fragment job).

It's also worth noting that the submit operation is atomic. In other
words, either all jobs are submitted, or none of them are (when an
error is detected when preparing jobs).

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>